NPC have their own script.
need to find NPC google docs to add the script
