+++ Assets/NPC.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c5b38e0117742c4458a0b96c77370903
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
