--- Assets/scripts/Chat/ChatUIBehavior.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5c13301901c3fa0409f498ac9d36c4aa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
