+++ Assets/scripts/Chat/DialogueTrigger.cs.meta
